+++ src/nodes/mod.rs
@@ -17,6 +17,8 @@ use smithay::wayland::shell::xdg::dialog::ToplevelDialogHint;
 mod dynamics;
 pub(crate) mod ipc;
 mod session_ops;
+mod keyboard_navigation;
+pub(crate) use keyboard_navigation::{transfer_window, pan_field};
 
 pub use ipc::handle_request;
 pub(crate) use ipc::move_selected_direction;